🚑️ Anchor 0.29.0 entrypoint lifetimes hack
Anchor 0.29.0 introduced more restrictive lifetimes in the entrypoint function that are not compatible with those used by previous Anchor versions and other native or SPL programs. This commit transmutes the entrypoint function type to be the same as the previous version. This seems to be working correctly, but it has to be considered in case the entrypoint invocation is changed in the future and the lifetime guarantees  of the entrypoint inputs would change.